move the facing check for aggro to Unit class
